frontend/requirements.txt
streamlit==1.53.1
pandas==2.2.2
oracledb==3.4.1
python-dotenv==1.0.1